Convert IndexedDBStore to TS
